[PyTorch Edge] Leave out field for debug_handle if not being built with eager symbolication support
Pull Request resolved: https://github.com/pytorch/pytorch/pull/66131

Turns out that a model with 72k instructions causes about 0.5MiB of additional memory overhead (if there's an 8 byte memory overhead per instruction). This is not necessary if we're building w/o eager symbolication support. This change eliminates the 8 byte `debug_handle` if the build is w/o eager symbolication support.
ghstack-source-id: 140045478

Differential Revision: [D31387784](https://our.internmc.facebook.com/intern/diff/D31387784/)